git-subtree is in git 1.7.11 releases.
People keep asking why, if git-subtree is not part of git, it's not in their
git.  Clarify the text to say why that is.

--

Hi,

This git-subtree repository is obsolete now that git-subtree has been merged
into mainline git (yay!).  If you have patches you want to make to
git-subtree, I encourage you to email them to the core git mailing list,
where the hordes of awesome maintainers are well known for responding very
quickly to queries and patches, very much unlike me.

Note: it is only available in git 1.7.11 and higher.  Also, it's in the
"contrib" subtree for now, so it's not installed by default.  Please
encourage your distribution packager (eg. Debian) to include contrib
when making their git packages.

Thanks!

Avery